+++ src/spoke/instances/SpokeInstance.sol
@@ -25,8 +25,8 @@ contract SpokeInstance is Spoke {
   /// @param maxUserBorrows The maximum allowed number of borrowed reserves per user.
   function initialize(
     address authority,
-    uint64 maxUserCollaterals,
-    uint64 maxUserBorrows
+    uint24 maxUserCollaterals,
+    uint24 maxUserBorrows
   ) external override reinitializer(SPOKE_REVISION) {
     emit UpdateOracle(ORACLE);
     require(authority != address(0), InvalidAddress());
